weather_alarm/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
discord.py>=2.3.0
python-telegram-bot>=20.7
loguru>=0.7.0
